Refactor Usage handling
 - Remove model specific Usage implementations
 - Add `Object getNativeUsage()` to Usage interface
   - This will allow the model specific Usage data to be returned
   - At the client side, client needs to cast the return type of `getNativeUsage` into the corresponding Usage returned by the model API
 - Rename `generationTokens` to `completionTokens`
   - Since `completion` token name is more common among the models, renaming generation tokens into completion tokens
 - Change the prompt, completion and total token return types to Integer
 - Use DefaultUsage for most of the model specific usage handling
    - When initializing set the native usage to the model specific usage type

Resolves #1407

+++ models/spring-ai-bedrock/src/main/java/org/springframework/ai/bedrock/anthropic3/BedrockAnthropic3ChatModel.java
@@ -132,8 +132,7 @@ public class BedrockAnthropic3ChatModel implements ChatModel, StreamingChatModel
 	}
 
 	protected Usage extractUsage(AnthropicChatResponse response) {
-		return new DefaultUsage(response.usage().inputTokens().longValue(),
-				response.usage().outputTokens().longValue());
+		return new DefaultUsage(response.usage().inputTokens(), response.usage().outputTokens());
 	}
 
 	/**
